Move all new files to refactor folder for ease of organization
