+++ tests/refactor/config/selectors/selectors.json
