+++ tests/refactor/devtools.spec.ts
@@ -1,0 +1,1 @@
+import {test, expect} from '@playwright/test';
